Scrape location and automate the coffee shop scraping
Add function to scrape physical address (location) of coffee shops.
Add 'location' to the outputted csv file for each coffee shop by updating the append_shop_to_csv function.
Update the format of adding coffee shops such that it is less manual work to add each coffee shop to the csv, using a shop_urls list and a loop.
Update the merging of CSV files to be automatic using a loop instead of manually calling the function for each shop.